Refactor Shoutbox to separate file
